+++ AutoTrader/Trader/Advisor/LeastSquareMethod.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -5,17 +6,17 @@ namespace AutoTrader.Advisor
 {
     public class LeastSquareMethod
     {
-        public static float Slope(IEnumerable<float> dataIn)
+        public static Decimal Slope(IEnumerable<Decimal> dataIn)
         {
 
             // Equation: m =  (n*sum(xy) - sum(x)*sum(y)) / (n*sum(x^2) - sum(x)^2)
             int n = dataIn.Count();
             int x = 1;
             int sumX = 0;
-            float sumY = 0;
-            float sumXY = 0;
-            float sumX2 = 0;
-            foreach (float y in dataIn)
+            Decimal sumY = 0;
+            Decimal sumXY = 0;
+            Decimal sumX2 = 0;
+            foreach (Decimal y in dataIn)
             {
                 sumX += x;
                 sumY += y;
