+++ AutoTrader/Trader/Repository/Lykke/PayloadDTOs/PayloadBalance.cs
@@ -5,8 +5,8 @@ namespace AutoTrader.Repository
     public class PayloadBalance
     {
         public String AssetId { get; set; }
-        public float Available { get; set; }
-        public float Reserved { get; set; }
+        public Decimal Available { get; set; }
+        public Decimal Reserved { get; set; }
         public float Timestamp { get; set; }
     }
 }
