+++ AutoTrader/Trader/Repository/Lykke/PayloadDTOs/PayloadPrice.cs
@@ -6,8 +6,8 @@ namespace AutoTrader.Repository
     public class PayloadPrice
     {
         public string assetPairId { get; set; }
-        public float bid { get; set; }
-        public float ask { get; set; }
+        public Decimal bid { get; set; }
+        public Decimal ask { get; set; }
 
         [JsonConverter(typeof(UnixMillisecondsConverter))]
         public DateTime timestamp { get; set; }
