+++ AutoTraderTests/Advisor/LeastSquareMethodTests.cs
@@ -1,6 +1,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using AutoTrader.Advisor;
+using System;
 
 namespace AutoTraderTests.Library
 {
@@ -11,16 +12,16 @@ namespace AutoTraderTests.Library
         [TestMethod()]
         public void SimpleLinearTest()
         {
-            List<float> y = new List<float> { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
-            float slope = LeastSquareMethod.Slope(y);
+            List<Decimal> y = new List<Decimal> { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+            Decimal slope = LeastSquareMethod.Slope(y);
             Assert.AreEqual(1, slope);
         }
 
         [TestMethod()]
         public void SimpleNegativeLinearTest()
         {
-            List<float> y = new List<float> { 5, 4, 3, 2, 1 };
-            float slope = LeastSquareMethod.Slope(y);
+            List<Decimal> y = new List<Decimal> { 5, 4, 3, 2, 1 };
+            Decimal slope = LeastSquareMethod.Slope(y);
             Assert.AreEqual(-1, slope);
         }
     }
